Protect the `/admin/panel` route
* Append the passport.authenticate() middleware to the login POST request
* Create a table & user for admin if doesn't exist
* Use an admin obj for storing env variables
* Adapt existent SQL queries to support the admin's table
* Implement the `verifyCb` to validate password

+++ controllers/adminController.js
@@ -2,11 +2,9 @@ const db = require('../database/queries');
 
 module.exports = {
     adminPanelGet: async (req, res) => {
-        const messages = await db.getAllMessages();
+        if (req.isAuthenticated()) {
+            const messages = await db.getAllMessages();
 
-        // TODO: find a way to pass a key-value pair for the redirect request
-        // if (req.login === 'granted') {
-        if (res) {
             res.render('adminPanel', {
                 title: 'Admin Panel',
                 messages,
@@ -16,13 +14,21 @@ module.exports = {
         }
     },
     deletePost: async (req, res) => {
-        const { postID } = req.params;
-        await db.deleteMessage(postID);
+        if (req.isAuthenticated()) {
+            const { postID } = req.params;
+            await db.deleteMessage(postID);
 
-        res.redirect('/admin/panel');
+            res.redirect('/admin/panel');
+        } else {
+            res.status(400).send('Access denied.');
+        }
     },
     deleteAllPosts: async (req, res) => {
-        await db.deleteAllMessages();
-        res.redirect('/admin/panel');
+        if (req.isAuthenticated()) {
+            await db.deleteAllRows();
+            res.redirect('/admin/panel');
+        } else {
+            res.status(400).send('Access denied.');
+        }
     },
 };
